+++ src/core/request/interface.ts
@@ -249,7 +249,7 @@ export interface RequestAPI {
  * Request engine
  */
 export interface RequestEngine {
-	(params: RequestOptions): AbortablePromise<Response>;
+	(request: RequestOptions, params: MiddlewareParams): AbortablePromise<Response>;
 
 	/**
 	 * The flag indicates that the active requests with the same request hash can be merged