+++ src/core/request/CHANGELOG.md
@@ -9,6 +9,16 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.??.?? (2022-??-??)
+
+#### :rocket: New Feature
+
+* Now `RequestEngine` will take `MiddlewareParams` as a second argument
+
+#### :bug: Bug Fix
+
+* Fixed a race condition while checking and saving pending requests to the cache
+
 ## v3.72.0 (2021-12-10)
 
 #### :rocket: New Feature